fix: move types to separate file
